# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Job": {
        "properties": {
          "changed_something": {
            "description": "False for a refusal: the store is untouched",
            "title": "Changed Something",
            "type": "boolean"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Recorded, never interpreted \u2014 `outcome` is the authority",
            "title": "Exit Code"
          },
          "gate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Which gate refused \u2014 branch on this, not the message",
            "title": "Gate"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Outcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "additionalProperties": true,
            "default": {},
            "title": "Result",
            "type": "object"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "verb": {
            "title": "Verb",
            "type": "string"
          }
        },
        "required": [
          "id",
          "verb",
          "entity_id",
          "state",
          "changed_something"
        ],
        "title": "Job",
        "type": "object"
      },
      "JobState": {
        "enum": [
          "queued",
          "running",
          "finished"
        ],
        "title": "JobState",
        "type": "string"
      },
...
      "NetWorth": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The date the view was COMPUTED at",
            "title": "As Of"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/EntityTotal"
            },
            "title": "Entities",
            "type": "array"
          },
          "granularity": {
            "const": "aggregate",
            "title": "Granularity",
            "type": "string"
          },
          "is_partial": {
            "description": "True when a declared entity could not be included",
            "title": "Is Partial",
            "type": "boolean"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "reporting_currency": {
            "title": "Reporting Currency",
            "type": "string"
          },
          "total": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Money"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "granularity",
          "reporting_currency",
          "is_partial",
          "entities",
          "provenance"
        ],
        "title": "NetWorth",
        "type": "object"
      },
      "Outcome": {
        "description": "WLC's own vocabulary, mirrored so callers never see a raw exit code.",
        "enum": [
          "ok",
          "attention",
          "refused",
          "failed"
        ],
        "title": "Outcome",
        "type": "string"
      },
...
      "Positions": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "excluded": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Excluded"
            },
            "title": "Excluded",
            "type": "array"
          },
          "granularity": {
            "const": "positions",
            "title": "Granularity",
            "type": "string"
          },
          "is_partial": {
            "title": "Is Partial",
            "type": "boolean"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "reporting_currency": {
            "title": "Reporting Currency",
            "type": "string"
          },
          "rows": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PositionRow"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "provenance",
          "granularity",
          "reporting_currency",
          "is_partial"
        ],
        "title": "Positions",
        "type": "object"
      },
...
      "Transactions": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "excluded": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Excluded"
            },
            "title": "Excluded",
            "type": "array"
          },
          "granularity": {
            "const": "transactions",
            "title": "Granularity",
            "type": "string"
          },
          "is_partial": {
            "title": "Is Partial",
            "type": "boolean"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "reporting_currency": {
            "title": "Reporting Currency",
            "type": "string"
          },
          "rows": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TransactionRow"
            },
            "title": "Rows",
            "type": "array"
          }
        },
        "required": [
          "provenance",
          "granularity",
          "reporting_currency",
          "is_partial"
        ],
        "title": "Transactions",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "WealthLens",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/jobs": {
      "get": {
        "description": "Everything this session has run. In memory by design (ADR-0002) \u2014 the UI says so.",
        "operationId": "list_jobs_api_jobs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "title": "Response List Jobs Api Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Jobs"
      },
      "post": {
        "description": "Run a WLC verb against exactly one named, manifest-declared workspace.\n\nThere is no \"run against all\": each store has its own gates, its own key and its own lock, and a\nfan-out would hide which one refused.",
        "operationId": "start_job_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Job"
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "operationId": "job_status_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Status"
      }
    },
    "/api/jobs/{job_id}/stream": {
      "get": {
        "description": "Watch a running verb as server-sent events.\n\nThe stream carries the verb's own narration as OPAQUE lines. Only `import` publishes structure\nmid-run, so parsing prose into a progress bar would mean inventing a contract that does not exist\nand breaking it on the next copy edit upstream. Lines now, structure when the engine offers it.\n\nThe stream is a convenience, never the record: the outcome stays retrievable from the job endpoint\nafter the stream closes, so a viewer who navigates away or reconnects still learns what happened.",
        "operationId": "job_stream_api_jobs__job_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Stream"
      }
    },
    "/api/networth": {
      "get": {
        "operationId": "networth_api_networth_get",
        "parameters": [
          {
            "in": "query",
            "name": "on",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "On"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetWorth"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Networth"
      }
    },
    "/api/positions": {
      "get": {
        "operationId": "positions_api_positions_get",
        "parameters": [
          {
            "in": "query",
            "name": "on",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "On"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Positions"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Positions"
      }
    },
    "/api/transactions": {
      "get": {
        "operationId": "transactions_api_transactions_get",
        "parameters": [
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transactions"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transactions"
      }
    },
    "/api/upload": {
      "post": {
        "description": "Deposit a statement into one entity's inbox \u2014 and do nothing else with it.\n\nNo parsing, no store write, no look inside. Importing is a separate, explicit act with WLC's gates\nin front of it (ADR-0005).",
        "operationId": "upload_api_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deposit"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload"
      }
    },
...
    "/api/workspace/{entity_id}": {
      "get": {
        "description": "Where an entity's money physically lives, what state it is in, and what built it.",
        "operationId": "workspace_detail_api_workspace__entity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workspace Detail"
      }
    },
    "/api/workspace/{entity_id}/settings": {
      "post": {
        "description": "Change one setting. Everything bootstrap asks is editable here (identity-and-settings).",
        "operationId": "change_settings_api_workspace__entity_id__settings_post",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Settings"
      }
    }
  }
}